+++ pygeos/geometry.py
@@ -132,7 +132,7 @@ def get_coordinate_dimension(geometry):
 def get_num_coordinates(geometry):
     """Returns the total number of coordinates in a geometry.
 
-    Returns -1 for not-a-geometry values.
+    Returns 0 for not-a-geometry values.
 
     Parameters
     ----------
@@ -147,7 +147,7 @@ def get_num_coordinates(geometry):
     >>> get_num_coordinates(Geometry("GEOMETRYCOLLECTION (POINT(0 0), LINESTRING(0 0, 1 1))"))
     3
     >>> get_num_coordinates(None)
-    -1
+    0
     """
     return lib.get_num_coordinates(geometry)
 
@@ -321,6 +321,8 @@ def get_point(geometry, index):
 def get_num_points(geometry):
     """Returns number of points in a linestring or linearring.
 
+    Returns 0 for not-a-geometry values.
+
     Parameters
     ----------
     geometry : Geometry or array_like
@@ -338,6 +340,8 @@ def get_num_points(geometry):
     >>> get_num_points(line)
     4
     >>> get_num_points(Geometry("MULTIPOINT (0 0, 1 1, 2 2, 3 3)"))
+    0
+    >>> get_num_points(None)
     0
     """
     return lib.get_num_points(geometry)
@@ -398,6 +402,8 @@ def get_interior_ring(geometry, index):
 def get_num_interior_rings(geometry):
     """Returns number of internal rings in a polygon
 
+    Returns 0 for not-a-geometry values.
+
     Parameters
     ----------
     geometry : Geometry or array_like
@@ -417,6 +423,8 @@ def get_num_interior_rings(geometry):
     >>> get_num_interior_rings(polygon_with_hole)
     1
     >>> get_num_interior_rings(Geometry("POINT (1 1)"))
+    0
+    >>> get_num_interior_rings(None)
     0
     """
     return lib.get_num_interior_rings(geometry)
@@ -506,6 +514,8 @@ def get_parts(geometry, return_index=False):
 def get_num_geometries(geometry):
     """Returns number of geometries in a collection.
 
+    Returns 0 for not-a-geometry values.
+
     Parameters
     ----------
     geometry : Geometry or array_like
@@ -523,5 +533,7 @@ def get_num_geometries(geometry):
     4
     >>> get_num_geometries(Geometry("POINT (1 1)"))
     1
+    >>> get_num_geometries(None)
+    0
     """
     return lib.get_num_geometries(geometry)